Making processors distinct files
