+++ docker/docker-compose.yml
@@ -86,6 +86,9 @@ services:
       NEO4J_DB_PORT: 7687
       NEO4J_USERNAME: "${NEO4J_USERNAME}"
       NEO4J_PASSWORD: "${NEO4J_PASSWORD}"
+    depends_on:
+      - kafka1
+      - weaviate
 
   event_processor:
     image: event-processor:0.1.1
@@ -97,6 +100,9 @@ services:
       NEO4J_DB_PORT: 7687
       NEO4J_USERNAME: "${NEO4J_USERNAME}"
       NEO4J_PASSWORD: "${NEO4J_PASSWORD}"
+    depends_on:
+      - kafka1
+      - neo4j
 
   document_processor:
     image: document-processor:0.1.1
@@ -107,6 +113,9 @@ services:
       GROQ_API_KEY: "${GROQ_API_KEY}"
       VECTOR_DB_HOST: "docker-weaviate-1"
       VECTOR_DB_PORT: 8080
+    depends_on:
+      - kafka1
+      - weaviate
   
   transcript_processor:
     image: transcript-processor:0.1.1
@@ -117,16 +126,22 @@ services:
       GROQ_API_KEY: "${GROQ_API_KEY}"
       VECTOR_DB_HOST: "docker-weaviate-1"
       VECTOR_DB_PORT: 8080
+    depends_on:
+      - kafka1
+      - weaviate
   
   slack_processor:
     image: slack-processor:0.1.1
     environment:
       KAFKA_BROKER: "kafka1:29092"
-      KAFKA_TOPIC: "transcripts"
+      KAFKA_TOPIC: "slack"
       PYTHONUNBUFFERED: 1
       GROQ_API_KEY: "${GROQ_API_KEY}"
       VECTOR_DB_HOST: "docker-weaviate-1"
       VECTOR_DB_PORT: 8080
+    depends_on:
+      - kafka1
+      - weaviate
       
   api:
     image: context-api:0.1.1
@@ -145,3 +160,7 @@ services:
       NEO4J_PASSWORD: "${NEO4J_PASSWORD}"
       GROQ_API_KEY: "${GROQ_API_KEY}"
       GSUITE_CREDS_FILE: "${GSUITE_CREDS_FILE}"
+    depends_on:
+      - kafka1
+      - weaviate
+      - neo4j
