+++ pyproject.toml
@@ -7,10 +7,10 @@ readme = "README.rst"
 packages = [{include = "library"}, {include = "processor"}]
 
 [tool.poetry.scripts]
-processor = "processor.main:start"
-event_processor = "processor.main:start_kafka_calendar"
-document_processor = "processor.main:start_kafka_documents"
-employee_processor = "processor.employees.main:start"
+processor = "processor.emails:start"
+event_processor = "processor.events:start_kafka_calendar"
+document_processor = "processor.documents:start_kafka_documents"
+employee_processor = "processor.employees:start"
 slack_processor = "processor.slack:start"
 transcripts_processor = "processor.transcripts:start"
